+++ src/eradiate/scenes/phase/_tabulated.py
@@ -61,26 +61,12 @@ class TabulatedPhaseFunction(PhaseFunction):
 
     Notes
     -----
-    The :math:`\mu` coordinate must cover the :math:`[-1, 1]` interval but
-    there is no constraint on value ordering or spacing. In particular,
-    irregular :math:`\mu` grids are supported.
+    * The :math:`\mu` coordinate must cover the :math:`[-1, 1]` interval but
+      there is no constraint on value ordering or spacing. In particular,
+      irregular :math:`\mu` grids are supported.
 
-    Since the underlying ``tabphase`` plugin expects phase function values on
-    a regular :math:`\mu` grid, this class will resample irregularly gridded
-    phase function data on a regular grid. The step of the resampling grid is
-    equal to the smallest :math:`\mu` step found in the input `data` field.
-
-    The resampling grid step selection policy ensures that the precision of
-    the input data is preserved, but the resampled data may be large (several
-    hundred MB) if no care is taken in the preparation of the input data.
-
-    To control the :math:`\mu` grid, you can simply provide a
-    :class:`~xarray.DataArray` object with a regularly gridded
-    :math:`\mu` coordinate ; in that case, the phase function data is not
-    resampled on the :math:`\mu` dimension.
-
-    For optimal performance, providing phase function data on a regular,
-    sorted :math:`\mu` grid is recommended.
+    * For optimal performance, providing phase function data on a regular,
+      sorted :math:`\mu` grid is recommended.
     """
 
     data: xr.DataArray = documented(
@@ -140,43 +126,14 @@ class TabulatedPhaseFunction(PhaseFunction):
         """
         w_units = self.data.w.attrs["units"]
 
-        # data already maps to regular grid of scattering angle cosines
-        dmu = self.data.mu.values[1:] - self.data.mu.values[:-1]
-        if np.allclose(dmu, dmu[0], rtol=1e-8):
-            return (
-                self.data.sel(i=0, j=0)
-                .interp(
-                    w=w.m_as(w_units),
-                    kwargs=dict(bounds_error=True),
-                )
-                .data
-            )
-
-        # data does not map to regular grid of scattering angle cosines
-        else:
-            # compute the regular grid from the smallest mu step found in the
-            # input data
-            dmu_min = np.abs(self.data.mu.diff(dim="mu")).values.min()
-            nmu = int(np.ceil(2.0 / dmu_min)) + 1
-            mu = np.linspace(-1.0, 1.0, nmu)
-
-            # interpolate first on wavelength
-            _w = _ensure_magnitude_array(w)
-            data_w_interpolated = self.data.sel(i=0, j=0).interp(
-                w=_w.m_as(w_units),
+        return (
+            self.data.isel(i=0, j=0)
+            .interp(
+                w=w.m_as(w_units),
                 kwargs=dict(bounds_error=True),
             )
-
-            # for performance, interpolation on mu dimension is performed with
-            # numpy.interp, which is found to be more than twice faster than
-            # xarray.DataArray.interp
-            phase = np.full(shape=(_w.size, nmu), fill_value=np.nan)
-            mup = self.data.mu.values
-            for iw in range(_w.size):
-                fp = data_w_interpolated.isel(w=iw).values
-                phase[iw, :] = np.interp(x=mu, xp=mup, fp=fp)
-
-            return phase.squeeze()
+            .data
+        )
 
     def eval_ckd(self, *bindexes: Bindex) -> np.ndarray:
         """
@@ -199,11 +156,31 @@ class TabulatedPhaseFunction(PhaseFunction):
         return self.eval_mono(w)
 
     def kernel_dict(self, ctx: KernelDictContext) -> KernelDict:
-        return KernelDict(
-            {
-                self.id: {
-                    "type": "tabphase",
-                    "values": ",".join(map(str, self.eval(ctx.spectral_ctx))),
+        # Evaluate phase function
+        phase_values = self.eval(ctx.spectral_ctx)
+
+        # Retrieve mu grid
+        mu = self.data.mu.values
+        dmu = mu[1:] - mu[:-1]
+
+        # Create kernel dict
+        if np.allclose(dmu, dmu[0]):  # mu grid is regular
+            return KernelDict(
+                {
+                    self.id: {
+                        "type": "tabphase",
+                        "values": ",".join(map(str, phase_values)),
+                    }
                 }
-            }
-        )
+            )
+
+        else:  # mu grid is irregular
+            return KernelDict(
+                {
+                    self.id: {
+                        "type": "tabphase_irregular",
+                        "values": ",".join(map(str, phase_values)),
+                        "nodes": ",".join(map(str, mu)),
+                    }
+                }
+            )